Refactor IssueLinkedIssues component into multiple files
The codebase has been refactored to split the complex IssueLinkedIssues component into separate, more manageable elements. This change improves readability and maintainability, as each subcomponent can now be understood and manipulated independently. The subcomponents were moved into their respective files: LinkIssueTypeDropdown, LinkIssueSearch, and the Index file for the IssueLinkedIssues.
